Minor Fixes
Improved the overall project's architecture
Improved meta-description generation
