+++ app/layout.js
@@ -4,8 +4,8 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "To do | NextJS",
+  description: "Consult todos, see their details and discover them by user !",
 };
 
 export default function RootLayout({ children }) {
