+++ app/page.js
@@ -1,13 +1,15 @@
 import styles from "./page.module.css";
-import Tasks from "./tasks/page.js";
+import Tasks from "./tasks/components/tasks.js";
+import getTasks from "./tasks/utils/getTasks.js";
 
-export default function Home() {
+export default async function Home() {
+  const tasks = await getTasks();
   return (
     <main className={styles.main}>
       <div className={styles.description}>
         <div className="taskList">
           <h1>My todo list</h1>
-          <Tasks />
+          <Tasks tasks={tasks} />
         </div>
       </div>
     </main>
