+++ app/tasks/user/[id]/page.js
@@ -1,19 +1,25 @@
-import Tasks from "../../page.js";
+import Link from "next/link.js";
+import Tasks from "../../components/tasks.js";
+import getTasksByUser from "../../utils/getTasksByUser.js";
 
-async function getTasksByUser(userId) {
-  const res = await fetch(
-    `https://jsonplaceholder.typicode.com/todos?userId=${userId}`,
-    {
-      cache: "no-store",
-    }
-  );
-  const tasks = await res.json();
-
-  return tasks;
+export async function generateMetadata({ params }) {
+  const tasks = await getTasksByUser(params.id);
+  const taskNum = tasks.length;
+  return {
+    title: `${params.id}'s tasks | ${taskNum}`,
+    description: `Discover all ${taskNum} ${params.id}'s tasks`,
+  };
 }
 
 export default async function UserTask({ params }) {
   const { id } = params;
   const userTasks = await getTasksByUser(id);
-  return <Tasks userTasks={userTasks} />;
+  return (
+    <>
+      <Link href={"/"}>See all tasks →</Link>
+      <br />
+      <br />
+      <Tasks tasks={userTasks} />
+    </>
+  );
 }
