+++ app/tasks/utils/getTasksByUser.js
@@ -1,0 +1,11 @@
+export default async function getTasksByUser(userId) {
+  const res = await fetch(
+    `https://jsonplaceholder.typicode.com/todos?userId=${userId}`,
+    {
+      cache: "no-store",
+    }
+  );
+  const tasks = await res.json();
+
+  return tasks;
+}
